requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
colorzero==2.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gpiozero==1.6.2
httplib2==0.20.2
idna==3.3
keyboard==0.13.5
oauthlib==3.1.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.7
wincertstore==0.2
